move SQLCheckResult to Common package
